Refactored CreateRecipe page into new component
